feat: create an accordion with a unique approach
create an interface that allows you to show and hide content by clicking on titles. Each item in the accordion has a question and an associated answer. When you click on the question, the corresponding answer appears below it. The idea is that only one answer can be shown at a time, to keep the interface clean and easy to use. If no data is available, it displays a message indicating this.